+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
+
+public abstract class Command {
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
+
+    public boolean isExit() {
+        return false;
+    }
+}
