+++ src/main/java/duke/command/ListCommand.java
@@ -1,0 +1,11 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
+
+public class ListCommand extends Command {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.printToFormat(tasks.toString());
+    }
+}
