+++ src/main/java/duke/commandtype/CommandType.java
@@ -1,0 +1,8 @@
+package duke.commandtype;
+
+/**
+ * a list of valid Commands as enum
+ */
+public enum CommandType {
+    BYE, LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT
+}
